create contours mod
