+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod cliffs;
 pub mod merge;
 pub mod util;
+pub mod contours;
